% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident_registry.R
\name{list_supported_incident_cities}
\alias{list_supported_incident_cities}
\title{List Supported Incident Cities}
\usage{
list_supported_incident_cities()
}
\value{
A tibble describing the supported city adapters.
}
\description{
Returns the cities currently supported by the standardized incident wrapper.
}
